chore: minor fixes and cleanup to .bin

+++ .bin/clean.sh
@@ -1,5 +1,7 @@
 main () {
-    rm -r build
+    if [ -d "build" ]; then
+        rm -r build
+    fi
 }
 
 main
